adding status

+++ server/controllers/dashboard/company.js
@@ -3,6 +3,7 @@ var common = require("../../../config/common");
 var Promise = require("promise");
 var DB = require("../../../config/database");
 var companyModel = require("../../models/companyModel");
+const employeeModel = require("../../models/employeeModel");
 
 async function postCompanyInfo(req, res, next) {
   var params = req.body;
@@ -76,8 +77,9 @@ async function postCompanyInfo(req, res, next) {
 }
 
 async function getList(req, res, next) {
+  const { status } = req.body;
   try {
-    const _company = await companyModel.getCompanyList();
+    const _company = await companyModel.getCompanyList(status);
     return common.send(res, 200, _company, "Success");
   } catch (err) {
     return common.send(res, 400, "", "Exception error: " + err);
@@ -176,22 +178,48 @@ async function deleteProduct(req, res, next) {
 
 async function deleteCompany(req, res, next) {
   var params = req.body;
-  const { index } = params;
-
+  const { index, status } = params;
+  var updated_at = moment(new Date()).format("YYYY-MM-DD hh:mm:ss");
   try {
-    let del_query = "DELETE FROM companies WHERE id = ? ";
-    let del_values = [index];
-
-    let del_result = await new Promise(function (resolve, reject) {
-      DB.query(del_query, del_values, function (err, data) {
-        if (err) reject(err);
-        else resolve(data.affectedRows > 0 ? true : false);
+    if( status === 1 ) {
+      let del_query = "DELETE FROM companies WHERE id = ? ";
+      let del_values = [index];
+  
+      let del_result = await new Promise(function (resolve, reject) {
+        DB.query(del_query, del_values, function (err, data) {
+          if (err) reject(err);
+          else resolve(data.affectedRows > 0 ? true : false);
+        });
       });
-    });
+      if (!del_result) return common.send(res, 300, "", "Database Error");
 
-    if (!del_result) return common.send(res, 300, "", "Database Error");
+    } else {
+      let _query =
+        "UPDATE companies SET status = ?, updated_at = ? WHERE id = ? ";
+      let _values = [1, updated_at, index];
 
-    const _company = await companyModel.getCompanyList();
+      let _result = await new Promise(function (resolve, reject) {
+        DB.query(_query, _values, function (err, data) {
+          if (err) reject(err);
+          else resolve(data.affectedRows > 0 ? true : false);
+        });
+      });
+      if (!_result) return common.send(res, 300, "", "Database Error");
+
+      let _emp_ = await employeeModel.findEmployeeByCompanyId(index);
+      let sendData = {
+        isLogout: true,
+      }
+      let receivers = [];
+      _emp_.forEach(e => {
+        receivers.push(e.push_token);
+      });
+      common.sendDataThroughFCMByType(receivers, sendData, 1, function (response) {
+        console.log('push data response => ', response);
+      })
+    }
+    
+    const _company = await companyModel.getCompanyList(status);
     return common.send(res, 200, _company, "Success");
   } catch (err) {
     return common.send(res, 400, "", "Exception error: " + err);
@@ -209,33 +237,53 @@ async function updateCompany(req, res, next) {
     biz_pin,
     tax,
     arr_employee,
+    status,
+    isUndo,
   } = params;
   var updated_at = moment(new Date()).format("YYYY-MM-DD hh:mm:ss");
 
   try {
-    let _query =
-      "UPDATE companies SET biz_name=?, biz_phone=?, biz_address=?, biz_email=?, biz_pin=?, tax=?, employee_ids=?, updated_at = ? WHERE id = ? ";
-    let _values = [
-      biz_name,
-      biz_phone,
-      biz_address,
-      biz_email,
-      biz_pin,
-      tax,
-      arr_employee.toString(),
-      updated_at,
-      company_id,
-    ];
+    if(isUndo) {
+      let _query =
+      "UPDATE companies SET status=?, updated_at = ? WHERE id = ? ";
+      let _values = [
+        0,
+        updated_at,
+        company_id,
+      ];
 
-    let _result = await new Promise(function (resolve, reject) {
-      DB.query(_query, _values, function (err, data) {
-        if (err) reject(err);
-        else resolve(data.affectedRows > 0 ? true : false);
+      let _result = await new Promise(function (resolve, reject) {
+        DB.query(_query, _values, function (err, data) {
+          if (err) reject(err);
+          else resolve(data.affectedRows > 0 ? true : false);
+        });
       });
-    });
-    if (!_result) return common.send(res, 300, "", "Database Error");
+      if (!_result) return common.send(res, 300, "", "Database Error");
+    } else {
+      let _query =
+      "UPDATE companies SET biz_name=?, biz_phone=?, biz_address=?, biz_email=?, biz_pin=?, tax=?, employee_ids=?, updated_at = ? WHERE id = ? ";
+      let _values = [
+        biz_name,
+        biz_phone,
+        biz_address,
+        biz_email,
+        biz_pin,
+        tax,
+        arr_employee.toString(),
+        updated_at,
+        company_id,
+      ];
 
-    const _company = await companyModel.getCompanyList();
+      let _result = await new Promise(function (resolve, reject) {
+        DB.query(_query, _values, function (err, data) {
+          if (err) reject(err);
+          else resolve(data.affectedRows > 0 ? true : false);
+        });
+      });
+      if (!_result) return common.send(res, 300, "", "Database Error");
+    }
+    
+    const _company = await companyModel.getCompanyList(status);
     return common.send(res, 200, _company, "Success");
   } catch (err) {
     return common.send(res, 400, "", "Exception error: " + err);
